Ajustando erros

+++ monitoramento/models.py
@@ -1,8 +1,8 @@
 from django.db import models
 
 class Leituras(models.Model):
-    temperatura = models.FloatField()
-    ph = models.IntegerField()
+    temperatura = models.IntegerField()
+    ph = models.FloatField()
     tds = models.IntegerField()
     data_hora = models.DateTimeField(auto_now_add=True)
